mv_ddr: change bus active mask to support 64 bit ecc
Reviewed-on: http://vgitil04.il.marvell.com:8080/31398
Tested-by: Star_Automation <star@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/31514
Tested-by: Nadav Haklai <nadavh@marvell.com>

+++ ddr_topology_def.h
@@ -192,7 +192,7 @@ struct mv_ddr_topology_map {
 	struct if_params interface_params[MAX_INTERFACE_NUM];
 
 	/* Bit mask for active buses */
-	u8 bus_act_mask;
+	u16 bus_act_mask;
 };
 
 /* DDR3 training global configuration parameters */
